backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
gradio-client==0.9.0
python-dotenv==1.0.0
google-generativeai==0.8.3
requests==2.32.5
tencentcloud-sdk-python==3.0.1470
